fix: model download Windows compatibility and build warnings
- Rewrite move_extracted_files to use copy+delete instead of rename
  (Windows cross-directory rename fails)
- Add copy_dir_recursive for nested directory handling
- Double flatten pass to handle multi-level nested tars
- Add post-extraction validation to catch incomplete downloads
- Fix clippy warnings (empty doc lines, unnecessary casts, etc)
- Add #![allow(unused)] to silence dead code warnings from
  App struct abstraction not fully utilized by main.rs
- Remove session_manager/memory_store dead code from App

+++ src/tts/synthesizer.rs
@@ -1,7 +1,6 @@
 /// TTS synthesizer using sherpa-onnx Kokoro
 ///
 /// Based on sherpa-onnx rust-api-examples/kokoro_tts_en.rs
-
 use sherpa_onnx::{
     GenerationConfig, OfflineTts, OfflineTtsConfig, OfflineTtsKokoroModelConfig,
 };
@@ -46,7 +45,7 @@ impl TtsSynthesizer {
         let tts = OfflineTts::create(&config)
             .ok_or_else(|| anyhow::anyhow!("Failed to create OfflineTts"))?;
 
-        let sample_rate = tts.sample_rate() as i32;
+        let sample_rate = tts.sample_rate();
         let num_speakers = tts.num_speakers();
 
         Ok(Self {
